requirements.txt
flask
pandas
mysql-connector-python
openpyxl
xlsxwriter
requests
